moved to a separate file manager_files_cli
