+++ lib/file_system/lib/constants/CMakeLists.txt
@@ -1,0 +1,4 @@
+add_library(constants src/constants.cpp)
+
+
+target_include_directories(constants PUBLIC include)
