+++ lib/file_system/lib/constants/include/constants.hpp
@@ -1,0 +1,7 @@
+#pragma once
+#include <string_view>
+
+namespace fs {
+inline constexpr uint16_t STANDARD_BUFFER_SIZE = 1024;
+inline constexpr std::string_view NAME_MAIN_DIR = ".main_dir";
+}
