+++ lib/file_system/lib/errors/CMakeLists.txt
@@ -1,0 +1,5 @@
+
+add_library(errors src/errors.cpp)
+
+
+target_include_directories(errors PUBLIC include)
