+++ lib/file_system/lib/errors/include/errors.hpp
@@ -1,0 +1,8 @@
+#pragma once
+#include <stdexcept>
+
+namespace fs {
+class FSError : public std::runtime_error {
+    using std::runtime_error::runtime_error;
+};
+}   // namespace fs
