+++ lib/file_system/lib/errors/src/errors.cpp
@@ -1,0 +1,1 @@
+#include "errors.hpp"
