+++ lib/file_system/lib/file/src/file.cpp
@@ -1,6 +1,6 @@
 #include "file.hpp"
 
-namespace file_for_fs {
+namespace file_fs {
 bool FID::operator<(const FID &other) const {
     if (hash_ < other.hash_) {
         return true;
@@ -13,4 +13,4 @@ bool FID::operator==(const FID &other) const { return hash_ == other.hash_; };
 
 std::string FID::string() const { return hash_; };
 
-}   // namespace file_for_fs
+}   // namespace file_fs
